Move git-store to version-control/git/controllers
