+++ src/version-control/git/views/git-view.tsx
@@ -1,9 +1,9 @@
 import { GitBranch, RefreshCw } from "lucide-react";
 import { useCallback, useEffect, useState } from "react";
 import { useBufferStore } from "@/stores/buffer-store";
-import { useGitStore } from "@/stores/git-store";
 import { cn } from "@/utils/cn";
 import { getBranches, getCommitDiff, getFileDiff, getGitLog, getGitStatus } from "@/utils/git";
+import { useGitStore } from "@/version-control/git/controllers/git-store";
 import type { MultiFileDiff } from "../../../components/diff-viewer/utils/types";
 
 // Import modular components
